rename source resource in api payload
